added url serach support and began working on displaying search results
