+++ IndieWritter/urls.py
@@ -38,7 +38,8 @@ urlpatterns = [
     path('<int:poem_id>/', local_poem_detail),
     re_path(r'profiles?/',include('profiles.urls')),
     path('api/poems/', include('poems.api.urls')),
-    path('api/search/', include ('search.urls')),
+    path('api/search/', include ('search.api.urls')),
+    path('search/', include('search.urls')),
     path('login', login_view),
     path('logout', logout_view),
     path('register', register_view),
