+++ indiewritter-web/public/index.html
@@ -44,7 +44,7 @@
     <div class ='' id="indiewritter" data-id = '192' data-username = 'kayvman' data-postprem = 'true'></div>
     <div class ='d-none' id="indiewritter-feed" data-id = '192' data-username = 'kayvman' data-postprem = 'true'></div>
  -->  
-    <div class =  'searchBar' id = 'searchBar' className = "px-1"></div>
+    <div class =  'searchBar' id = 'searchResult' className = "px-1"></div>
     
     <!--
       This HTML file is a template.
